real-life calculator simulator
It works just like a regular calculator, but it was made because of some interesting html properties for input, such as "data-number". Which can lter be accessed in js using "dataset.number". This is the fundamental concept of this project. There is also an additional concept, and that is - in terms of number data, the datas need to be evaluated first (eval) then displayed (screen.value). :)